gcloud-reader/requirements.txt
flask==3.1.0
google-cloud-texttospeech==2.21.1
python-dotenv==1.0.1
gunicorn==23.0.0
